src/task/requirements.txt
amqp==2.5.2
billiard==3.6.3.0
CacheControl==0.12.6
cachetools==4.1.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
firebase-admin==4.2.0
Flask==1.1.2
google-api-core==1.17.0
google-api-python-client==1.8.3
google-auth==1.14.3
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.28.1
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
httplib2==0.17.3
idna==2.9
importlib-metadata==1.6.0
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.8
MarkupSafe==1.1.1
matchbox-orm==0.2.7
msgpack==1.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
redis==3.5.1
requests==2.23.0
rsa==4.0
six==1.14.0
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
Werkzeug==1.0.1
zipp==3.1.0
